Introduced param handling and client decoupling
