+++ DCL_PiXYZ/DCL_PiXYZ.csproj
@@ -30,6 +30,7 @@
     </Target>
 
     <ItemGroup>
+      <PackageReference Include="CommandLineParser" Version="2.9.1" />
       <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
       <PackageReference Include="PiXYZCSharpAPI" Version="2024.1.0.17-win64" />
       <PackageReference Include="SharpGLTF.Core" Version="1.0.0-alpha0031" />
